Parse command if it can't parse it as a function.
Ok, so quick note. I've been testing this on the /etc/profile provided
with Fedora 34 WS. This starts with two comment blocks, then a function,
containing a case, containing an if, containing a command. But there's a
problem.

If I allow for support of parsing the command, due to a bug I have yet
to diagnose, it can't parse the function. I suspect this is because of
how `attempt` works; it attempts to parse the function as a command.